requirements_client.txt
replicate>=0.25.0
pillow>=9.0.0
